Elementary Add Friends function
Using updated form in index.ejs, user can input friend's username they want to add. If friend is not in database, nothing changes. If the proposed friend is in the database, that user's ID will be added to 'notorBIG's friend array in MongoDB. For now can only add friends to user notorBIG, but functionality to add friends and check that friend exists is implemented.