With new implemented api this is adding the first template of the "forms" we will usee to retrieve any data from the client and send to our server/mongodb.
The form I have created for now is a song form where the user can add
whatever title/writer and hit send. This will then save an id,
songName, and writer for the song and store it in the songPost schema.

The posts files will be used to display any songs/other data we want
from the backend to the client. For example if we wanted to display a
prompt we could create a PromptPost file and then display from the
PromptPost schema. This is the intended functionality but I have not
fully implemented

My songForm along with basic front end will need some styling, reviewed
with jake and the login files can be easily implemented as a type of
form. As you can see in my songForm file I used very similar approach
to what you guys were doing.