Edit footer for navigation
In this commit, I changed the footer. Now the footer can be used to
navigate from page to page, just like the navigation bar at the top. I
also made the names hyperlinks to each person's respective Github
account. I also edited some of the other files so that they look better
with the footer. This mostly included adding a bottom margin to the
lowest element