Follow friends feature implemented with database
Logged in users can add friends through the form at the bottom of home.js. It updates the friends array for whichever user is logged in (must be in 'users' collection). Returns error messages if the user is already following that user or that user is not in our 'users' collection.