Added access to user profile data
When on the home page, 'userdata' in localStorage should hold the user's data if they're logged in (so it's available to get the current users songs, friends, etc). The 'getplaylists' function returns an array of playlists to the console holding all the user's playlists, can be used to print on profile. 'setUser' sets the 'userdata' to the user logged in. 'newPlaylist' loops through each of the user's friends and compiles a playlist of each friend's current song for the day and returns it to the console.