Create playlists for each user based on friend responses
New function in userFetchBack called compileResponses. Using the song data in tempSongs database that holds user, prompt, and their song response, this function gathers all the data from the users database and uses that to parse through all the data in tempSongs to find responses only by the users friends. It then groups the songs based on prompt, so the users playlists are objects of their friends responses to the same prompt. I tried to make it as independed as I could from the collection documents to allow for updates to the type of song data being stored.
Also changed the userData model so the playlists are now arrays of objects, not strings.